agents/commodity_agent/requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
requests==2.31.0
numpy==1.25.2
pandas==2.1.4
scipy==1.11.4

# Multi-Tool Integration Dependencies
# LangChain Integration
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-core==0.1.10
langchain-experimental==0.0.49

# LlamaIndex Integration
llama-index==0.9.48
llama-index-embeddings-openai==0.1.6
llama-index-llms-openai==0.1.12
llama-index-vector-stores-postgres==0.1.4

# Haystack Integration
farm-haystack==1.21.2
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# AutoGen Integration
pyautogen==0.2.14
openai==1.6.1

# Computer Use Integration (if available)
# computer-use==0.1.0  # Uncomment when available

# Additional Dependencies
python-multipart==0.0.6
httpx==0.25.2
websockets==12.0
redis==5.0.1
celery==5.3.4 